requirements.txt
streamlit
diffusers
transformers
accelerate
safetensors
torch
Pillow
numpy
opencv-python
scipy
